Incorporate my cut-down GNU gzip into zgz and use it in pristine-gz
With this change, zgz now provides support for explicitly setting the
timestamp and original filename in both GNU gzip and zlib gzip.  Use
this support in pristine-gz to do all compression with -c, eliminating
the complexity caused by gzip removing files after compressing them.

The new pristine-gz requires the new zgz and vice-versa, but deltas
generated with the new pristine-gz do not depend on the new zgz.