pristine-tar: Fix behavior in the corner case where the tarball puts all files in a subdir, but does not contain the subdir's directory entry.
I think that this block of code has always been buggy. It attempts to
see if all files listed in the manifest are present inside the subdir.
Usually, though, the manifest *includes* the subdir, and so the code
checked to see if $subdir/$subdir existed, and concluded everything was not
in the subdir.

Turns out there is one tar file that lists files inside a subdir, but
does not include the subdir. In that case, the code concludes correctly
that all files are in a subdir. But, the result of that turns out to be
bad; unsettting $subdir breaks the build since it then tries to move the
wrong path.