pristine-tar: handle properly Perl gzips
Perl's Compress::Raw::Zlib interfaces directly with zlib and
apparently is the only implementation out there which tunes a very
specific parameter of zlib, memLevel, to 9, instead of 8 which is
the default.

The module is used, among others, by Compress::Gzip which in turn is
used by IO::Zlib.

It was found on the real world on tarballs generated by Perl 5.10's
Module::Build (cf. #618284)