add --quirk suse, which affects --old-bzip2 mode to inexactly replicate suse bzip2
Suse's bzip2 is modifed for some reason to use a huffman table of 17 bytes.

While suse has a newer bzip2 than the version in old-bzip2, it was most
expedient to keep only one copy of bzip2 in zgz, and parameterize it to
emulate suse's patched newer version. This is done by using the blocksort
code from bzip 1.0.5.

This does not 100% replicate the output of suse's bzip2, but the
differences seem to just be in some preliminary header info, and are
quite small. So this will be combined with using xdelta.