pristine-tar: Fail when the delta is excessively large, probably due to the tarball being compressed with something that tar x can  auto-extract, but that pristine-tar does not support.
It would be cleaner to test the tarball's magic to see if it's a regular,
non-compressed tarball. But, there are a lot of historical tar-like
formats, so implementing magic for tar is hard. (And I don't use file(1)
in pristine-tar).

Anyway, this also catches any other failure that can result in a delta >=
the size of the input. Completely failing to get any of the input files
right could also make this happen, perhaps.