Factorize tar and xdelta program names into distinct variables at the beginning of the pristine-tar script.
Pristine-tar uses options (e.g., "--owner"), which are only present in GNU tar.
Pristine-tar uses the old 1.1 release series Xdelta, which is syntactically incompatible to the 3.0 release series of Xdelta.

Therefore, this commit centralizes the definition of the program names of tar and xdelta so that their exact names can be easily configured.