Factorize external program names into variables
The external programs (tar, xdelta) called by the pristine-tar script have
been hardcoded until now.  This requires that both program names are the
first in the user's PATH, which can lead to problems, especially on
non-GNU/Linux platforms:

 * Pristine-tar uses options (e.g., "--owner"), which are only present
   in GNU tar.
 * Pristine-tar uses the old 1.1 release series Xdelta, which is
   syntactically incompatible to the 3.0 release series of Xdelta.

This commit centralizes the definition of the program names of tar and
xdelta so that their exact names (or locations)  can be easily configured.

Signed-off-by: Philipp A. Hartmann <philipp.hartmann@offis.de>