Don't use inplace editing option '-i' of sed for version number extraction from debian/changelog to update pristine-tar.spec.
The -i option is not supported by all versions of sed (e.g. not by SunOS 5.9 nor by NetBSD 6.0). Additionally, the exact syntax may vary. E.g., the sed of Mac OS X wants -i to be followed by a suffix, which is added to the original filename to create a backup file. The portable solution is to avoid the in-place editing alltogether and to instead write first to a temporary file, which is then moved back to the original filename.