Fix suse-bzip2 Makefile to compile a dynamic libbz2.dylib on Mac OS X and ensure that zgz sets DYLD_LIBRARY_PATH on Mac OS X to swap in a different library.
The suse-bzip2 Makefile was only able to handle the compilation of shared object libraries like they are used on Linux. Mac OS X uses dynamic libraries, which require different compiler and linker switches. Additionally, DYLD_LIBRARY_PATH has to be set instead of LD_LIBRARY_PATH to find dynamic libraries in a non-standard path. Therefore, the zgz source needed a small patch as well.

For the future, consider to rewrite the Makefiles to be less platform-specific. E.g., use GNU libtool to compile static and dynamic libraries with the right flags on different platforms. Another approach could be to avoid the swapping in of dynamic libraries altogether to change the behavior of  a certain archiver and to provide statically-linked archiver binaries.