Improve configuration code for the Perl scripts in Makefile.PL to make the source files valid Perl scripts.
The pristine-tar Perl scripts are now generated from pristine-tar.pl, pristine-gz.pl, pristine-bz2.pl, and pristine-xz.pl instead of from pristine-tar.in, pristine-gz.in, pristine-bz2.in, and pristine-xz.in, respectively. As now the assignment to $tar_program and $xdelta_program is matched instead of the @TAR_PROGRAM@ and @XDELTA_PROGRAM@ placeholders, the *.pl source files are valid Perl scripts.

This patch was suggested by Philipp A. Hartmann <philipp.hartmann@offis.de>.