Makefile.PL: add message about custom parameters
Since adding custom parameters to Makefile.PL scripts is not
transparently supported by MakeMaker, recent versions print a
note about unknown options given to the script.

Nevertheless, such parameters are handled correctly and are
passed through to a reinvocation in case of a changed Makefile.PL.
A message is printed to inform the user about this fact.

Signed-off-by: Philipp A. Hartmann <philipp.hartmann@offis.de>