Makefile.PL: add custom build parameters
This patch introduces three custom build parameters

 * PERL_SHEBANG
     Configure the #! line in the installed perl scripts.
     May be needed on some passwords, in case MakeMaker
     does not get the installation right
 * TAR_PROGRAM, XDELTA_PROGRAM
     Set a custom name (or location) for the required (GNU) tar and
     xdelta programs, in case they are not callable through their
     default names from the $PATH

If set, the variables are used to munge the corresponding parts of
the scripts.  The $tar_program and $xdelta_program variables in
pristine-tar are now marked as such.

The patch itself extends the "macro" section of the MakeMaker generated
Makefile.  To call the inherited function, the package needs to be set
explicitly.

This patch is an updated version, incorporating a spacing fix when
using multiple replacements, as discovered by Torsten Maehne during
his review.

Signed-off-by: Philipp A. Hartmann <philipp.hartmann@offis.de>
Reviewed-by: Torsten Maehne <Torsten.Maehne@gmx.de>